requirements.txt
python-telegram-bot>=21.0,<22.0
pydantic>=2.7,<3.0
pypdf>=4.2,<6.0
reportlab>=4.2,<5.0
python-dotenv>=1.0,<2.0
google-genai>=1.30.0,<2.0
PyMuPDF>=1.24
